reorganization of the main processor. More modularization. More support files
